requirements.txt
opencv-python==4.3.0.36
numpy==1.18.5
aruco==3.1.2.0
ar-markers==0.5.0
paho-mqtt==1.5.0
opencv-contrib-python==4.5.1.48
pyyaml==3.12
